# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "CanonicalLabel": {
        "description": "Canonical PII category. Every detector's raw output maps into this taxonomy.",
        "enum": [
          "PERSON",
          "EMAIL",
          "PHONE",
          "ADDRESS",
          "URL",
          "DATE",
          "ACCOUNT",
          "SECRET",
          "USERNAME",
          "DEMOGRAPHIC",
          "ORGANIZATION",
          "OCCUPATION",
          "MONEY",
          "VEHICLE",
          "PHYSICAL"
        ],
        "title": "CanonicalLabel",
        "type": "string"
      },
...
      "DetectorOptions": {
        "description": "Per-detector options namespace. Each key carries options for that one\ndetector. Only the entry matching the top-level `detector` is consulted \u2014\nother keys are accepted but ignored, so a client can carry one options\nblob across detector swaps without restructuring it.",
        "properties": {
          "opf": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OpfOptions"
              },
              {
                "type": "null"
              }
            ],
            "description": "Options for the `opf` detector. Ignored by other detectors."
          }
        },
        "title": "DetectorOptions",
        "type": "object"
      },
...
      "FindRequest": {
        "description": "Base request for `/api/find` \u2014 no text-rewriting fields.",
        "examples": [
          {
            "categories": [
              "EMAIL"
            ],
            "detector": "presidio",
            "text": "Email joe@example.com about the trip to Elgin, TX."
          }
        ],
        "properties": {
          "categories": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CanonicalLabel"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Canonical categories to keep. Omit or set to `null` to keep every category the detector produces. An empty list `[]` is a deliberate \"match nothing\" filter and returns zero spans \u2014 use `null` if you mean \"no filter\". Values outside the canonical taxonomy are rejected with `422`.",
            "examples": [
              [
                "EMAIL",
                "PHONE"
              ]
            ],
            "title": "Categories"
          },
          "detector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Detector name from `GET /api/detectors`. Omit to use the server's `DEFAULT_DETECTOR` (set via env, typically `opf`).",
            "examples": [
              "presidio",
              "opf",
              "gliner"
            ],
            "title": "Detector"
          },
          "options": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DetectorOptions"
              },
              {
                "type": "null"
              }
            ],
            "description": "Per-detector options namespace. The entry matching the top-level `detector` is consulted; other entries are accepted but ignored.",
            "examples": [
              {
                "opf": {
                  "decode_mode": "argmax"
                }
              }
            ]
          },
          "text": {
            "description": "Free-form text to scan for PII.",
            "examples": [
              "Email joe@example.com about the trip to Elgin, TX."
            ],
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "FindRequest",
        "type": "object"
      },
      "FindResponse": {
        "description": "Response from `/api/find`.",
        "examples": [
          {
            "detected_spans": [
              {
                "end": 21,
                "label": "EMAIL",
                "raw_label": "EMAIL_ADDRESS",
                "start": 6,
                "text": "joe@example.com"
              }
            ],
            "detector": "presidio",
            "summary": {
              "by_label": {
                "EMAIL": 1
              },
              "span_count": 1
            },
            "text": "Email joe@example.com about the trip to Elgin, TX."
          }
        ],
        "properties": {
          "detected_spans": {
            "description": "Spans detected by the chosen detector.",
            "items": {
              "$ref": "#/components/schemas/SpanOut"
            },
            "title": "Detected Spans",
            "type": "array"
          },
          "detector": {
            "description": "Detector that produced the spans.",
            "title": "Detector",
            "type": "string"
          },
          "summary": {
            "$ref": "#/components/schemas/SummaryOut"
          },
          "text": {
            "description": "Echo of the request `text`.",
            "title": "Text",
            "type": "string"
          },
          "warning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Non-fatal warning surfaced by the detector backend, if any.",
            "title": "Warning"
          }
        },
        "required": [
          "detector",
          "text",
          "detected_spans",
          "summary"
        ],
        "title": "FindResponse",
        "type": "object"
      },
...
      "ReplaceRequest": {
        "description": "Request for `/api/replace` \u2014 `FindRequest` plus a `mode` field that picks\nhow detected spans are rewritten.",
        "examples": [
          {
            "detector": "presidio",
            "mode": "label_token",
            "text": "Email alice@x.com or call +1-415-555-0100."
          }
        ],
        "properties": {
          "categories": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CanonicalLabel"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Canonical categories to keep. Omit or set to `null` to keep every category the detector produces. An empty list `[]` is a deliberate \"match nothing\" filter and returns zero spans \u2014 use `null` if you mean \"no filter\". Values outside the canonical taxonomy are rejected with `422`.",
            "examples": [
              [
                "EMAIL",
                "PHONE"
              ]
            ],
            "title": "Categories"
          },
          "detector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Detector name from `GET /api/detectors`. Omit to use the server's `DEFAULT_DETECTOR` (set via env, typically `opf`).",
            "examples": [
              "presidio",
              "opf",
              "gliner"
            ],
            "title": "Detector"
          },
          "mode": {
            "default": "label",
            "description": "How to rewrite each detected span. `redact` -> `********` (fixed asterisks). `label` -> `[EMAIL]` (default). `label_number` -> `[EMAIL_1]` (per-request counter; duplicates reuse the number). `label_token` -> `[EMAIL_jRc7QGn]` (deterministic 7-char Skyflow vault token; requires `SKYFLOW_TOKEN_VAULT_*` env).",
            "enum": [
              "redact",
              "label",
              "label_number",
              "label_token"
            ],
            "examples": [
              "label",
              "label_token"
            ],
            "title": "Mode",
            "type": "string"
          },
          "options": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DetectorOptions"
              },
              {
                "type": "null"
              }
            ],
            "description": "Per-detector options namespace. The entry matching the top-level `detector` is consulted; other entries are accepted but ignored.",
            "examples": [
              {
                "opf": {
                  "decode_mode": "argmax"
                }
              }
            ]
          },
          "text": {
            "description": "Free-form text to scan for PII.",
            "examples": [
              "Email joe@example.com about the trip to Elgin, TX."
            ],
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "ReplaceRequest",
        "type": "object"
      },
      "ReplaceResponse": {
        "description": "Response from `/api/replace`.",
        "examples": [
          {
            "detected_spans": [
              {
                "end": 17,
                "label": "EMAIL",
                "raw_label": "EMAIL_ADDRESS",
                "replacement": "[EMAIL_MGaE1Bo]",
                "start": 6,
                "text": "alice@x.com"
              },
              {
                "end": 42,
                "label": "PHONE",
                "raw_label": "PHONE_NUMBER",
                "replacement": "[PHONE_vRXiWKZ]",
                "start": 27,
                "text": "+1-415-555-0100"
              }
            ],
            "detector": "presidio",
            "mode": "label_token",
            "replaced_text": "Email [EMAIL_MGaE1Bo] or call [PHONE_vRXiWKZ].",
            "summary": {
              "by_label": {
                "EMAIL": 1,
                "PHONE": 1
              },
              "span_count": 2
            },
            "text": "Email alice@x.com or call +1-415-555-0100."
          }
        ],
        "properties": {
          "detected_spans": {
            "description": "Spans detected by the chosen detector, with their replacements.",
            "items": {
              "$ref": "#/components/schemas/ReplacedSpan"
            },
            "title": "Detected Spans",
            "type": "array"
          },
          "detector": {
            "description": "Detector that produced the spans.",
            "examples": [
              "presidio"
            ],
            "title": "Detector",
            "type": "string"
          },
          "mode": {
            "description": "Echo of the request `mode`.",
            "enum": [
              "redact",
              "label",
              "label_number",
              "label_token"
            ],
            "examples": [
              "label_token"
            ],
            "title": "Mode",
            "type": "string"
          },
          "replaced_text": {
            "description": "`text` with each detected span replaced by its `replacement`. Overlapping spans: the earlier-starting span wins; later overlaps are skipped here (they still appear in `detected_spans`).",
            "examples": [
              "Email [EMAIL_MGaE1Bo]."
            ],
            "title": "Replaced Text",
            "type": "string"
          },
          "summary": {
            "$ref": "#/components/schemas/SummaryOut"
          },
          "text": {
            "description": "Echo of the request `text`.",
            "examples": [
              "Email alice@x.com."
            ],
            "title": "Text",
            "type": "string"
          },
          "warning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Non-fatal warning surfaced by the detector backend, if any.",
            "title": "Warning"
          }
        },
        "required": [
          "detector",
          "mode",
          "text",
          "detected_spans",
          "replaced_text",
          "summary"
        ],
        "title": "ReplaceResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "name": "local-privacy",
      "url": "https://github.com/jstjoe/local-privacy"
    },
...
    "summary": "Unified PII detection across OPF, GLiNER, Presidio, and Skyflow.",
    "title": "Privacy-detection API",
    "version": "0.5.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/find": {
      "post": {
        "description": "Run the chosen detector over `text` and return canonical-labelled spans.\n\nNo rewriting is performed \u2014 call `/api/replace` for that.\n\nFilter detector output to a subset of canonical labels with `categories`.\nOPF-only: pass `options.opf.decode_mode` to override the default Viterbi decoding.",
        "operationId": "find_api_find_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindResponse"
                }
              }
            },
            "description": "Detected spans plus per-label counts."
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "label_token_unconfigured": {
                    "summary": "label_token mode without vault env",
                    "value": {
                      "detail": "mode='label_token' requires SKYFLOW_TOKEN_VAULT_URL and SKYFLOW_TOKEN_VAULT_ID env vars to be set"
                    }
                  },
                  "unknown_detector": {
                    "summary": "Unknown detector",
                    "value": {
                      "detail": "unknown detector 'foo'; available: ['gliner', 'opf', 'presidio']"
                    }
                  }
                }
              }
            },
            "description": "Unknown detector or missing config for the chosen mode."
          },
          "422": {
            "content": {
              "application/json": {
                "examples": {
                  "bad_category": {
                    "summary": "Non-canonical value in `categories`",
                    "value": {
                      "detail": [
                        {
                          "input": "FOO",
                          "loc": [
                            "body",
                            "categories",
                            0
                          ],
                          "msg": "Input should be 'PERSON', 'EMAIL', 'PHONE', ...",
                          "type": "enum"
                        }
                      ]
                    }
                  },
                  "unknown_opf_option": {
                    "summary": "Unknown key in `options.opf`",
                    "value": {
                      "detail": [
                        {
                          "input": "argmax",
                          "loc": [
                            "body",
                            "options",
                            "opf",
                            "decod_mode"
                          ],
                          "msg": "Extra inputs are not permitted",
                          "type": "extra_forbidden"
                        }
                      ]
                    }
                  }
                }
              }
            },
            "description": "Request body failed Pydantic validation. Common causes: a value in `categories` that isn't a canonical label, an unknown key in `options.opf`, or a wrong type on any field."
          },
          "502": {
            "content": {
              "application/json": {
                "examples": {
                  "detector_failure": {
                    "summary": "Detector backend errored",
                    "value": {
                      "detail": "skyflow: upstream 503"
                    }
                  },
                  "vault_failure": {
                    "summary": "Token vault call failed",
                    "value": {
                      "detail": "label_token: vault insert failed: 401"
                    }
                  }
                }
              }
            },
            "description": "Detector backend failure or vault call failure."
          }
        },
        "summary": "Find sensitive data in text",
        "tags": [
          "Find"
        ]
      }
    },
...
    "/api/replace": {
      "post": {
        "description": "Find spans, then rewrite each one under the chosen `mode`. Four modes, in increasing strength of identity preservation:\n\n| `mode` | Looks like | What it preserves |\n|---|---|---|\n| `redact` | `********` | Nothing \u2014 fixed 8-char asterisk run regardless of span length. |\n| `label` | `[EMAIL]` | Category only. Default. |\n| `label_number` | `[EMAIL_1]` | Identity **within one request** via per-label counter; duplicate `(label, text)` reuses its number. |\n| `label_token` | `[EMAIL_MGaE1Bo]` | Identity **across requests and detectors** via a Skyflow vault. Deterministic \u2014 same plaintext maps to the same 7-char token forever. |\n\n**Overlapping spans:** the earlier-starting span wins; later overlaps are skipped in `replaced_text` (they still appear in `detected_spans`).\n\n**`label_token` requirements:** `SKYFLOW_TOKEN_VAULT_URL`, `SKYFLOW_TOKEN_VAULT_ID`, and a bearer (`SKYFLOW_TOKEN_BEARER_TOKEN`, falling back to `SKYFLOW_BEARER_TOKEN`). The vault must be configured per the token-vault setup guide \u2014 one table with one `tok_<label>` column per canonical label, each `DETERMINISTIC_FPT` with regex `^[A-Za-z0-9]{7}$`. Spans whose canonical label has no vault column fall back to `[LABEL]` for that span only.",
        "operationId": "replace_api_replace_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplaceResponse"
                }
              }
            },
            "description": "Replaced text, the spans that were rewritten, and per-label counts."
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "label_token_unconfigured": {
                    "summary": "label_token mode without vault env",
                    "value": {
                      "detail": "mode='label_token' requires SKYFLOW_TOKEN_VAULT_URL and SKYFLOW_TOKEN_VAULT_ID env vars to be set"
                    }
                  },
                  "unknown_detector": {
                    "summary": "Unknown detector",
                    "value": {
                      "detail": "unknown detector 'foo'; available: ['gliner', 'opf', 'presidio']"
                    }
                  }
                }
              }
            },
            "description": "Unknown detector or missing config for the chosen mode."
          },
          "422": {
            "content": {
              "application/json": {
                "examples": {
                  "bad_category": {
                    "summary": "Non-canonical value in `categories`",
                    "value": {
                      "detail": [
                        {
                          "input": "FOO",
                          "loc": [
                            "body",
                            "categories",
                            0
                          ],
                          "msg": "Input should be 'PERSON', 'EMAIL', 'PHONE', ...",
                          "type": "enum"
                        }
                      ]
                    }
                  },
                  "unknown_opf_option": {
                    "summary": "Unknown key in `options.opf`",
                    "value": {
                      "detail": [
                        {
                          "input": "argmax",
                          "loc": [
                            "body",
                            "options",
                            "opf",
                            "decod_mode"
                          ],
                          "msg": "Extra inputs are not permitted",
                          "type": "extra_forbidden"
                        }
                      ]
                    }
                  }
                }
              }
            },
            "description": "Request body failed Pydantic validation. Common causes: a value in `categories` that isn't a canonical label, an unknown key in `options.opf`, or a wrong type on any field."
          },
          "502": {
            "content": {
              "application/json": {
                "examples": {
                  "detector_failure": {
                    "summary": "Detector backend errored",
                    "value": {
                      "detail": "skyflow: upstream 503"
                    }
                  },
                  "vault_failure": {
                    "summary": "Token vault call failed",
                    "value": {
                      "detail": "label_token: vault insert failed: 401"
                    }
                  }
                }
              }
            },
            "description": "Detector backend failure or vault call failure."
          }
        },
        "summary": "Replace sensitive data in text",
        "tags": [
          "Replace"
        ]
      }
    }
  },
  "servers": [
    {
      "description": "Local dev",
      "url": "http://localhost:8000"
    }
  ],
...
}